+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/factory/ProcessingTypeDataProviderFactory.scala
@@ -2,7 +2,6 @@ package pl.touk.nussknacker.ui.factory
 
 import akka.actor.ActorSystem
 import pl.touk.nussknacker.engine.{CombinedProcessingTypeData, ConfigWithUnresolvedVersion, ProcessingTypeData}
-import pl.touk.nussknacker.ui.process.ProcessCategoryService
 import pl.touk.nussknacker.ui.process.deployment.DeploymentService
 import pl.touk.nussknacker.ui.process.processingtypedata.ProcessingTypeDataProvider
 import sttp.client3.SttpBackend
@@ -14,8 +13,7 @@ trait ProcessingTypeDataProviderFactory {
 
   def create(
       designerConfig: ConfigWithUnresolvedVersion,
-      deploymentServiceSupplier: Supplier[DeploymentService],
-      categoriesService: ProcessCategoryService
+      deploymentServiceSupplier: Supplier[DeploymentService]
   )(
       implicit ec: ExecutionContext,
       actorSystem: ActorSystem,