+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/factory/ProcessingTypeDataReaderBasedProcessingTypeDataProviderFactory.scala
@@ -2,7 +2,6 @@ package pl.touk.nussknacker.ui.factory
 
 import akka.actor.ActorSystem
 import pl.touk.nussknacker.engine.{CombinedProcessingTypeData, ConfigWithUnresolvedVersion, ProcessingTypeData}
-import pl.touk.nussknacker.ui.process.ProcessCategoryService
 import pl.touk.nussknacker.ui.process.deployment.DeploymentService
 import pl.touk.nussknacker.ui.process.processingtypedata.{ProcessingTypeDataProvider, ProcessingTypeDataReader}
 import sttp.client3.SttpBackend
@@ -14,15 +13,13 @@ object ProcessingTypeDataReaderBasedProcessingTypeDataProviderFactory extends Pr
 
   override def create(
       designerConfig: ConfigWithUnresolvedVersion,
-      deploymentServiceSupplier: Supplier[DeploymentService],
-      categoriesService: ProcessCategoryService
+      deploymentServiceSupplier: Supplier[DeploymentService]
   )(
       implicit ec: ExecutionContext,
       actorSystem: ActorSystem,
       sttpBackend: SttpBackend[Future, Any]
   ): ProcessingTypeDataProvider[ProcessingTypeData, CombinedProcessingTypeData] = {
-    implicit val deploymentService: DeploymentService         = deploymentServiceSupplier.get()
-    implicit val categoriesServiceImp: ProcessCategoryService = categoriesService
+    implicit val deploymentService: DeploymentService = deploymentServiceSupplier.get()
     ProcessingTypeDataReader.loadProcessingTypeData(designerConfig)
   }